requirements.txt
mysql-connector-python>=9.2.0
loguru>=0.7.3
pandas>=2.2.3